Refactor to satisfy swiftlint
