+++ .swiftlint.yml
@@ -1,0 +1,43 @@
+excluded:
+  - Example/Pods
+
+included:
+  - Gizmo
+
+disabled_rules:
+ - trailing_whitespace
+
+force_try: warning
+force_cast: warning
+line_length: 
+  warning: 200
+  error: 300
+  ignores_urls: true
+  ignores_function_declarations: false
+  ignores_comments: true
+  ignores_interpolated_strings: false
+file_length:
+  warning: 400
+  error: 500
+type_name:
+  min_length:
+    warning: 3
+    error: 2
+  max_length:
+    warning: 50
+    error: 60
+identifier_name:
+  min_length:
+    error: 3
+  excluded:
+    - id
+    - url
+    - int
+    - am
+    - pm
+large_tuple:
+   - 3
+   - 4
+function_parameter_count:
+  warning: 5
+  error: 6
