+++ Gizmo/Classes/Common/TimeZone/TimeZone+Africa.swift
@@ -1,0 +1,68 @@
+//
+//  TimeZone+Africa.swift
+//  Gizmo
+//
+//  Created by Nayanda Haberty on 28/9/22.
+//
+
+import Foundation
+
+public extension TimeZone {
+    
+    // MARK: Africa
+    
+    enum Africa: String, CaseIterable {
+        case abidjan = "Africa/Abidjan"
+        case accra = "Africa/Accra"
+        case addisAbaba = "Africa/Addis_Ababa"
+        case algiers = "Africa/Algiers"
+        case asmara = "Africa/Asmara"
+        case bamako = "Africa/Bamako"
+        case bangui = "Africa/Bangui"
+        case banjul = "Africa/Banjul"
+        case bissau = "Africa/Bissau"
+        case blantyre = "Africa/Blantyre"
+        case brazzaville = "Africa/Brazzaville"
+        case bujumbura = "Africa/Bujumbura"
+        case cairo = "Africa/Cairo"
+        case casablanca = "Africa/Casablanca"
+        case ceuta = "Africa/Ceuta"
+        case conakry = "Africa/Conakry"
+        case dakar = "Africa/Dakar"
+        case darEsSalaam = "Africa/Dar_es_Salaam"
+        case djibouti = "Africa/Djibouti"
+        case douala = "Africa/Douala"
+        case elAaiun = "Africa/El_Aaiun"
+        case freetown = "Africa/Freetown"
+        case gaborone = "Africa/Gaborone"
+        case harare = "Africa/Harare"
+        case johannesburg = "Africa/Johannesburg"
+        case juba = "Africa/Juba"
+        case kampala = "Africa/Kampala"
+        case khartoum = "Africa/Khartoum"
+        case kigali = "Africa/Kigali"
+        case kinshasa = "Africa/Kinshasa"
+        case lagos = "Africa/Lagos"
+        case libreville = "Africa/Libreville"
+        case lome = "Africa/Lome"
+        case luanda = "Africa/Luanda"
+        case lubumbashi = "Africa/Lubumbashi"
+        case lusaka = "Africa/Lusaka"
+        case malabo = "Africa/Malabo"
+        case maputo = "Africa/Maputo"
+        case maseru = "Africa/Maseru"
+        case mbabane = "Africa/Mbabane"
+        case mogadishu = "Africa/Mogadishu"
+        case monrovia = "Africa/Monrovia"
+        case nairobi = "Africa/Nairobi"
+        case ndjamena = "Africa/Ndjamena"
+        case niamey = "Africa/Niamey"
+        case nouakchott = "Africa/Nouakchott"
+        case ouagadougou = "Africa/Ouagadougou"
+        case portoNovo = "Africa/Porto-Novo"
+        case saoTome = "Africa/Sao_Tome"
+        case tripoli = "Africa/Tripoli"
+        case tunis = "Africa/Tunis"
+        case windhoek = "Africa/Windhoek"
+    }
+}
