+++ Gizmo/Classes/Common/TimeZone/TimeZone+Asia.swift
@@ -1,0 +1,120 @@
+//
+//  TimeZone+Asia.swift
+//  Gizmo
+//
+//  Created by Nayanda Haberty on 28/9/22.
+//
+
+import Foundation
+
+public extension TimeZone {
+    
+    // MARK: Asia
+    
+    enum Asia: String, CaseIterable {
+        case aden = "Asia/Aden"
+        case almaty = "Asia/Almaty"
+        case amman = "Asia/Amman"
+        case anadyr = "Asia/Anadyr"
+        case aqtau = "Asia/Aqtau"
+        case aqtobe = "Asia/Aqtobe"
+        case ashgabat = "Asia/Ashgabat"
+        case atyrau = "Asia/Atyrau"
+        case baghdad = "Asia/Baghdad"
+        case bahrain = "Asia/Bahrain"
+        case baku = "Asia/Baku"
+        case bangkok = "Asia/Bangkok"
+        case barnaul = "Asia/Barnaul"
+        case beirut = "Asia/Beirut"
+        case bishkek = "Asia/Bishkek"
+        case brunei = "Asia/Brunei"
+        case calcutta = "Asia/Calcutta"
+        case chita = "Asia/Chita"
+        case choibalsan = "Asia/Choibalsan"
+        case chongqing = "Asia/Chongqing"
+        case colombo = "Asia/Colombo"
+        case damascus = "Asia/Damascus"
+        case dhaka = "Asia/Dhaka"
+        case dili = "Asia/Dili"
+        case dubai = "Asia/Dubai"
+        case dushanbe = "Asia/Dushanbe"
+        case famagusta = "Asia/Famagusta"
+        case gaza = "Asia/Gaza"
+        case harbin = "Asia/Harbin"
+        case hebron = "Asia/Hebron"
+        case hoChiMinh = "Asia/Ho_Chi_Minh"
+        case hongKong = "Asia/Hong_Kong"
+        case hovd = "Asia/Hovd"
+        case irkutsk = "Asia/Irkutsk"
+        case jakarta = "Asia/Jakarta"
+        case jayapura = "Asia/Jayapura"
+        case jerusalem = "Asia/Jerusalem"
+        case kabul = "Asia/Kabul"
+        case kamchatka = "Asia/Kamchatka"
+        case karachi = "Asia/Karachi"
+        case kashgar = "Asia/Kashgar"
+        case kathmandu = "Asia/Kathmandu"
+        case katmandu = "Asia/Katmandu"
+        case khandyga = "Asia/Khandyga"
+        case krasnoyarsk = "Asia/Krasnoyarsk"
+        case kualaLumpur = "Asia/Kuala_Lumpur"
+        case kuching = "Asia/Kuching"
+        case kuwait = "Asia/Kuwait"
+        case macau = "Asia/Macau"
+        case magadan = "Asia/Magadan"
+        case makassar = "Asia/Makassar"
+        case manila = "Asia/Manila"
+        case muscat = "Asia/Muscat"
+        case nicosia = "Asia/Nicosia"
+        case novokuznetsk = "Asia/Novokuznetsk"
+        case novosibirsk = "Asia/Novosibirsk"
+        case omsk = "Asia/Omsk"
+        case oral = "Asia/Oral"
+        case phnomPenh = "Asia/Phnom_Penh"
+        case pontianak = "Asia/Pontianak"
+        case pyongyang = "Asia/Pyongyang"
+        case qatar = "Asia/Qatar"
+        case qostanay = "Asia/Qostanay"
+        case qyzylorda = "Asia/Qyzylorda"
+        case rangoon = "Asia/Rangoon"
+        case riyadh = "Asia/Riyadh"
+        case sakhalin = "Asia/Sakhalin"
+        case samarkand = "Asia/Samarkand"
+        case seoul = "Asia/Seoul"
+        case shanghai = "Asia/Shanghai"
+        case singapore = "Asia/Singapore"
+        case srednekolymsk = "Asia/Srednekolymsk"
+        case taipei = "Asia/Taipei"
+        case tashkent = "Asia/Tashkent"
+        case tbilisi = "Asia/Tbilisi"
+        case tehran = "Asia/Tehran"
+        case thimphu = "Asia/Thimphu"
+        case tokyo = "Asia/Tokyo"
+        case tomsk = "Asia/Tomsk"
+        case ulaanbaatar = "Asia/Ulaanbaatar"
+        case urumqi = "Asia/Urumqi"
+        case ustNera = "Asia/Ust-Nera"
+        case vientiane = "Asia/Vientiane"
+        case vladivostok = "Asia/Vladivostok"
+        case yakutsk = "Asia/Yakutsk"
+        case yangon = "Asia/Yangon"
+        case yekaterinburg = "Asia/Yekaterinburg"
+        case yerevan = "Asia/Yerevan"
+    }
+    
+    // MARK: Indian
+    
+    enum Indian: String, CaseIterable {
+        case antananarivo = "Indian/Antananarivo"
+        case chagos = "Indian/Chagos"
+        case christmas = "Indian/Christmas"
+        case cocos = "Indian/Cocos"
+        case comoro = "Indian/Comoro"
+        case kerguelen = "Indian/Kerguelen"
+        case mahe = "Indian/Mahe"
+        case maldives = "Indian/Maldives"
+        case mauritius = "Indian/Mauritius"
+        case mayotte = "Indian/Mayotte"
+        case reunion = "Indian/Reunion"
+    }
+}
