+++ Gizmo/Classes/Common/TimeZone/TimeZone+Europe.swift
@@ -1,0 +1,76 @@
+//
+//  TimeZone+Europe.swift
+//  Gizmo
+//
+//  Created by Nayanda Haberty on 28/9/22.
+//
+
+import Foundation
+
+public extension TimeZone {
+    
+    // MARK: Europe
+    
+    enum Europe: String, CaseIterable {
+        case amsterdam = "Europe/Amsterdam"
+        case andorra = "Europe/Andorra"
+        case astrakhan = "Europe/Astrakhan"
+        case athens = "Europe/Athens"
+        case belgrade = "Europe/Belgrade"
+        case berlin = "Europe/Berlin"
+        case bratislava = "Europe/Bratislava"
+        case brussels = "Europe/Brussels"
+        case bucharest = "Europe/Bucharest"
+        case budapest = "Europe/Budapest"
+        case busingen = "Europe/Busingen"
+        case chisinau = "Europe/Chisinau"
+        case copenhagen = "Europe/Copenhagen"
+        case dublin = "Europe/Dublin"
+        case gibraltar = "Europe/Gibraltar"
+        case guernsey = "Europe/Guernsey"
+        case helsinki = "Europe/Helsinki"
+        case isleOfMan = "Europe/Isle_of_Man"
+        case istanbul = "Europe/Istanbul"
+        case jersey = "Europe/Jersey"
+        case kaliningrad = "Europe/Kaliningrad"
+        case kiev = "Europe/Kiev"
+        case kirov = "Europe/Kirov"
+        case lisbon = "Europe/Lisbon"
+        case ljubljana = "Europe/Ljubljana"
+        case london = "Europe/London"
+        case luxembourg = "Europe/Luxembourg"
+        case madrid = "Europe/Madrid"
+        case malta = "Europe/Malta"
+        case mariehamn = "Europe/Mariehamn"
+        case minsk = "Europe/Minsk"
+        case monaco = "Europe/Monaco"
+        case moscow = "Europe/Moscow"
+        case oslo = "Europe/Oslo"
+        case paris = "Europe/Paris"
+        case podgorica = "Europe/Podgorica"
+        case prague = "Europe/Prague"
+        case riga = "Europe/Riga"
+        case rome = "Europe/Rome"
+        case samara = "Europe/Samara"
+        case sanMarino = "Europe/San_Marino"
+        case sarajevo = "Europe/Sarajevo"
+        case saratov = "Europe/Saratov"
+        case simferopol = "Europe/Simferopol"
+        case skopje = "Europe/Skopje"
+        case sofia = "Europe/Sofia"
+        case stockholm = "Europe/Stockholm"
+        case tallinn = "Europe/Tallinn"
+        case tirane = "Europe/Tirane"
+        case ulyanovsk = "Europe/Ulyanovsk"
+        case uzhgorod = "Europe/Uzhgorod"
+        case vaduz = "Europe/Vaduz"
+        case vatican = "Europe/Vatican"
+        case vienna = "Europe/Vienna"
+        case vilnius = "Europe/Vilnius"
+        case volgograd = "Europe/Volgograd"
+        case warsaw = "Europe/Warsaw"
+        case zagreb = "Europe/Zagreb"
+        case zaporozhye = "Europe/Zaporozhye"
+        case zurich = "Europe/Zurich"
+    }
+}
