+++ Gizmo/Classes/Common/TimeZone/TimeZone+Oceania.swift
@@ -1,0 +1,75 @@
+//
+//  TimeZone+Oceania.swift
+//  Gizmo
+//
+//  Created by Nayanda Haberty on 28/9/22.
+//
+
+import Foundation
+
+public extension TimeZone {
+    
+    // MARK: Australia
+    
+    enum Australia: String, CaseIterable {
+        case adelaide = "Australia/Adelaide"
+        case brisbane = "Australia/Brisbane"
+        case brokenHill = "Australia/Broken_Hill"
+        case currie = "Australia/Currie"
+        case darwin = "Australia/Darwin"
+        case eucla = "Australia/Eucla"
+        case hobart = "Australia/Hobart"
+        case lindeman = "Australia/Lindeman"
+        case lordHowe = "Australia/Lord_Howe"
+        case melbourne = "Australia/Melbourne"
+        case perth = "Australia/Perth"
+        case sydney = "Australia/Sydney"
+    }
+    
+    // MARK: Pacific
+    
+    enum Pacific: String, CaseIterable {
+        case apia = "Pacific/Apia"
+        case auckland = "Pacific/Auckland"
+        case bougainville = "Pacific/Bougainville"
+        case chatham = "Pacific/Chatham"
+        case chuuk = "Pacific/Chuuk"
+        case easter = "Pacific/Easter"
+        case efate = "Pacific/Efate"
+        case enderbury = "Pacific/Enderbury"
+        case fakaofo = "Pacific/Fakaofo"
+        case fiji = "Pacific/Fiji"
+        case funafuti = "Pacific/Funafuti"
+        case galapagos = "Pacific/Galapagos"
+        case gambier = "Pacific/Gambier"
+        case guadalcanal = "Pacific/Guadalcanal"
+        case guam = "Pacific/Guam"
+        case honolulu = "Pacific/Honolulu"
+        case johnston = "Pacific/Johnston"
+        case kanton = "Pacific/Kanton"
+        case kiritimati = "Pacific/Kiritimati"
+        case kosrae = "Pacific/Kosrae"
+        case kwajalein = "Pacific/Kwajalein"
+        case majuro = "Pacific/Majuro"
+        case marquesas = "Pacific/Marquesas"
+        case midway = "Pacific/Midway"
+        case nauru = "Pacific/Nauru"
+        case niue = "Pacific/Niue"
+        case norfolk = "Pacific/Norfolk"
+        case noumea = "Pacific/Noumea"
+        case pagoPago = "Pacific/Pago_Pago"
+        case palau = "Pacific/Palau"
+        case pitcairn = "Pacific/Pitcairn"
+        case pohnpei = "Pacific/Pohnpei"
+        case ponape = "Pacific/Ponape"
+        case portMoresby = "Pacific/Port_Moresby"
+        case rarotonga = "Pacific/Rarotonga"
+        case saipan = "Pacific/Saipan"
+        case tahiti = "Pacific/Tahiti"
+        case tarawa = "Pacific/Tarawa"
+        case tongatapu = "Pacific/Tongatapu"
+        case truk = "Pacific/Truk"
+        case wake = "Pacific/Wake"
+        case wallis = "Pacific/Wallis"
+    }
+}
