+++ Gizmo/Classes/UIKit/CGSize+Extensions.swift
@@ -56,42 +56,42 @@ public extension CGSize {
 
 // MARK: Operator
 
-@inlinable public func +(_ lhs: CGSize, _ rhs: CGSize) -> CGSize {
+@inlinable public func + (_ lhs: CGSize, _ rhs: CGSize) -> CGSize {
     CGSize(
         width: lhs.width + rhs.width,
         height: lhs.height + rhs.height
     )
 }
 
-@inlinable public func -(_ lhs: CGSize, _ rhs: CGSize) -> CGSize {
+@inlinable public func - (_ lhs: CGSize, _ rhs: CGSize) -> CGSize {
     CGSize(
         width: lhs.width - rhs.width,
         height: lhs.height - rhs.height
     )
 }
 
-@inlinable public func +(_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
+@inlinable public func + (_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
     CGSize(
         width: lhs.width + rhs,
         height: lhs.height + rhs
     )
 }
 
-@inlinable public func -(_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
+@inlinable public func - (_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
     CGSize(
         width: lhs.width - rhs,
         height: lhs.height - rhs
     )
 }
 
-@inlinable public func *(_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
+@inlinable public func * (_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
     CGSize(
         width: lhs.width * rhs,
         height: lhs.height * rhs
     )
 }
 
-@inlinable public func /(_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
+@inlinable public func / (_ lhs: CGSize, _ rhs: CGFloat) -> CGSize {
     CGSize(
         width: lhs.width / rhs,
         height: lhs.height / rhs
